# api/openapi.yaml
# Generated with protoc-gen-openapi
# https://github.com/google/gnostic/tree/master/cmd/protoc-gen-openapi

openapi: 3.0.3
info:
    title: ""
    version: 0.0.1
paths:
    /v1/app/creatapp:
        post:
            tags:
                - App
            operationId: App_CreatApp
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/http.app.CreatAppRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/http.app.CreatAppReply'
    /v1/app/getappbyid:
        post:
            tags:
                - App
            operationId: App_GetAppById
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/http.app.GetAppByIdRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/http.app.GetAppByIdReply'
    /v1/app/listapppage:
        post:
            tags:
                - App
            description: 获取应用列表分页
            operationId: App_ListAppPage
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/http.app.ListAppPageRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/http.app.ListAppPageReply'
    /v1/app/modifyapp:
        post:
            tags:
                - App
            operationId: App_ModifyApp
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/http.app.ModifyAppRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/http.app.ModifyAppReply'
    /v1/post/createpost:
        post:
            tags:
                - Post
            description: 创建帖子
            operationId: Post_CreatePost
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/http.post.CreatePostRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/http.post.CreatePostReply'
    /v1/user/getuser:
        post:
            tags:
                - User
            operationId: User_GetCurrentUser
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/http.user.GetCurrentUserRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/http.user.GetCurrentUserReply'
    /v1/user/list:
        post:
            tags:
                - User
            operationId: User_List
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/http.user.ListRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/http.user.ListReply'
    /v1/user/listbypage:
        post:
            tags:
                - User
            operationId: User_ListUserByPage
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/http.user.ListUserByPageRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/http.user.ListUserByPageReply'
    /v1/user/login:
        post:
            tags:
                - User
            description: 用户端登录
            operationId: User_Login
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/http.user.LoginRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/http.user.LoginReply'
    /v1/user/register:
        post:
            tags:
                - User
            description: 用户注册
            operationId: User_Register
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/http.user.RegisterRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/http.user.RegisterReply'
components:
    schemas:
        http.app.AppInfo:
            type: object
            properties:
                appName:
                    type: string
                appDesc:
                    type: string
                appIcon:
                    type: string
                appType:
                    type: integer
                    format: int32
                scoringStrategy:
                    type: integer
                    format: int32
                reviewStatus:
                    type: integer
                    format: int32
                reviewMessage:
                    type: string
                reviewId:
                    type: string
                userId:
                    type: string
                appId:
                    type: string
        http.app.AppModInfo:
            type: object
            properties:
                appName:
                    type: string
                appDesc:
                    type: string
                appIcon:
                    type: string
                appType:
                    type: integer
                    format: int32
                scoringStrategy:
                    type: integer
                    format: int32
        http.app.CreatAppReply:
            type: object
            properties: {}
        http.app.CreatAppRequest:
            type: object
            properties:
                appName:
                    type: string
                appDesc:
                    type: string
                appIcon:
                    type: string
                appType:
                    type: integer
                    format: int32
                scoringStrategy:
                    type: integer
                    format: int32
        http.app.GetAppByIdReply:
            type: object
            properties:
                app:
                    $ref: '#/components/schemas/http.app.AppInfo'
                user:
                    $ref: '#/components/schemas/http.app.UserInfo'
        http.app.GetAppByIdRequest:
            type: object
            properties:
                appId:
                    type: string
        http.app.ListAppPageReply:
            type: object
            properties:
                appList:
                    type: array
                    items:
                        $ref: '#/components/schemas/http.app.AppInfo'
                total:
                    type: integer
                    format: int32
        http.app.ListAppPageRequest:
            type: object
            properties:
                current:
                    type: integer
                    format: int32
                pageSize:
                    type: integer
                    format: int32
        http.app.ModifyAppReply:
            type: object
            properties: {}
        http.app.ModifyAppRequest:
            type: object
            properties:
                appId:
                    type: string
                appModInfo:
                    $ref: '#/components/schemas/http.app.AppModInfo'
        http.app.UserInfo:
            type: object
            properties:
                id:
                    type: string
                account:
                    type: string
                password:
                    type: string
                username:
                    type: string
                avatarUrl:
                    type: string
                gender:
                    type: integer
                    format: int32
                userInfo:
                    type: string
                userRole:
                    type: integer
                    format: int32
        http.post.CreatePostReply:
            type: object
            properties:
                id:
                    type: string
        http.post.CreatePostRequest:
            type: object
            properties:
                title:
                    type: string
                text:
                    type: string
                tags:
                    type: string
                userId:
                    type: string
        http.user.GetCurrentUserReply:
            type: object
            properties:
                id:
                    type: string
                account:
                    type: string
                password:
                    type: string
                username:
                    type: string
                avatarUrl:
                    type: string
                gender:
                    type: integer
                    format: int32
                userInfo:
                    type: string
                userRole:
                    type: integer
                    format: int32
        http.user.GetCurrentUserRequest:
            type: object
            properties: {}
        http.user.ListReply:
            type: object
            properties:
                username:
                    type: string
                avatarUrl:
                    type: string
                userInfo:
                    type: string
        http.user.ListRequest:
            type: object
            properties:
                id:
                    type: string
        http.user.ListUserByPageReply:
            type: object
            properties:
                userList:
                    type: array
                    items:
                        $ref: '#/components/schemas/http.user.UserInfo'
                total:
                    type: integer
                    format: int32
        http.user.ListUserByPageRequest:
            type: object
            properties:
                current:
                    type: integer
                    format: int32
                pageSize:
                    type: integer
                    format: int32
        http.user.LoginReply:
            type: object
            properties:
                id:
                    type: string
                token:
                    type: string
            description: 返回前端的请求
        http.user.LoginRequest:
            type: object
            properties:
                account:
                    type: string
                password:
                    type: string
            description: |-
                Register 用户登录
                 前端传来的请求
        http.user.RegisterReply:
            type: object
            properties:
                id:
                    type: string
            description: 返回前端的请求
        http.user.RegisterRequest:
            type: object
            properties:
                account:
                    type: string
                password:
                    type: string
                checkPassword:
                    type: string
            description: |-
                Register 用户注册
                 前端传来的请求
        http.user.UserInfo:
            type: object
            properties:
                id:
                    type: string
                account:
                    type: string
                password:
                    type: string
                username:
                    type: string
                avatarUrl:
                    type: string
                gender:
                    type: integer
                    format: int32
                userInfo:
                    type: string
                userRole:
                    type: integer
                    format: int32
tags:
    - name: App
    - name: Post
    - name: User